requirements.txt
# --extra-index-url https://download.pytorch.org/whl/cu118

decord==0.6.0
# deepspeed==0.11.1
einops==0.7.0
hydra-core==1.3.2
lightning==2.1.0
lightning-bolts==0.7.0
opencv-python-headless==4.8.1.78
pandas==2.0.3
scikit-learn==1.3.1
tensorboard==2.14.1
tensorboardx==2.6.2.2
torchmetrics==1.2.0
pytorchvideo==0.1.5
# torch==2.0.1+cu118
# torchvision==0.15.2+cu118
